Add DAO, repository, database Add formActivity, listActivity Save Estate and EstateImage

+++ app/src/main/java/com/openclassrooms/realestatemanager/data/model/Status.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "status")
 data class Status(
-        @PrimaryKey(autoGenerate = true) val id: Int = 0,
+        @PrimaryKey(autoGenerate = true) val id: Long = 0,
         @ColumnInfo(name = "name") val name: String
 )
